Dedupe AST shebang-insertion code into helper function
