+++ src/parser-typescript.js
@@ -1,6 +1,7 @@
 "use strict";
 
 const createError = require("./parser-create-error");
+const includeShebang = require("./parser-include-shebang");
 
 function parse(text) {
   const jsx = isProbablyJsx(text);
@@ -18,30 +19,7 @@ function parse(text) {
   }
 
   delete ast.tokens;
-
-  if (!text.startsWith("#!")) {
-    return ast;
-  }
-
-  const index = text.indexOf("\n");
-  const shebang = text.slice(2, index);
-  const comment = {
-    type: "Line",
-    value: shebang,
-    range: [0, index],
-    loc: {
-      start: {
-        line: 1,
-        column: 0
-      },
-      end: {
-        line: 1,
-        column: index
-      }
-    }
-  };
-  ast.comments = [comment].concat(ast.comments);
-
+  includeShebang(text, ast);
   return ast;
 }
